refactor(scripting): separate editor conversion and script execution
Split editor action state and display formatting, event projection, script
reading and writing, tree expansion, and runtime execution into focused units.
Share typed screen-reading syntax and input parsing across editor, CLI, and
runtime consumers while retaining persisted macro compatibility.

Keep conversion, parsing, and execution regression tests with their owners.

+++ src/CrossMacro.Infrastructure/Services/Editing/EditorActionProjectionValidator.cs
@@ -185,7 +185,7 @@ internal sealed class EditorActionProjectionValidator(IEditorActionConverter val
             return (false, ValidationMessages.ScrollAmountCannotBeZero);
         }
 
-        if (Math.Abs(action.ScrollAmount) > EditorActionValidationLimits.MaxScrollAmount)
+        if (Math.Abs((long)action.ScrollAmount) > EditorActionValidationLimits.MaxScrollAmount)
         {
             return (false, ValidationMessages.ScrollAmountTooLarge);
         }
